requirements.txt
fastapi==0.105.0
uvicorn==0.24.0.post1
Pillow==10.1.0
requests==2.31.0